fix fn+insert sleep
This commit fixes the key 'fn + insert' to sleep. When you active fn + insert, the computer will get into sleep but wake after a while and then it continue to sleep when the screen going black. If you can not wake the laptop by activating any key, you can try to active 'fn + insert' key again, it always effect. If you have other issues about it, you can open it and I will work on it in my free time.